+++ src/pg-promise/index.js
@@ -1,0 +1,3 @@
+const { vector, registerType } = require('../pg');
+
+module.exports = {vector, registerType};