+++ types/pg-promise/index.d.ts
@@ -1,0 +1,3 @@
+import { vector } from "../pg";
+import { registerType } from "../pg";
+export { vector, registerType };